requirements.txt
openai-whisper
numpy
torch
tqdm
more-itertools
tiktoken
